Flip month-match expressions in VI ng(à)y regex
Reverse the order of month digit matching, which greedily captures
months starting with 1, but only of a single digit. (10, 11, and 12) are
left out of this calculation, and thus regexes for capturing ng(à)y
fail.

Reversing will attempt to capture a two digit month beginning with 1,
then fall back to single digit months optionally preceeded with a 0.

Resolve #225.