Current:
if the fractional hour expression describes the hour fraction with term
like "quarter or quarters", then duckling couldn't correctly recognize it.

Expected:

Duckling should be able to identify this kind of expression and parse it
correctly.

Fix:

Add new rule to parse the fractional hour pattern that contains the
keyword like "quarter or quarters".