Ignore the no-"-1" linter error for duckling
Summary:
By default, Facebook Haskell code has a lint error banning
-1s, because at one point it was used as a default value in some
API handlers and it's better to use Option + Nothing for this use case.

But in Duckling - particularly in the Time module - it's quite
common that we actually want to work with -1 as a meaningful integer
value, involving some kind of offset.

Reviewed By: chessai

Differential Revision: D27118984

fbshipit-source-id: 5fe2200e8005a20855d7fdd3a8eb2ad33291edc8